+++ backend/pokerogue/src/main/java/com/pokerogue/helper/battle/service/TypeMultiplierProvider.java
@@ -3,7 +3,7 @@ package com.pokerogue.helper.battle.service;
 import com.pokerogue.helper.global.exception.ErrorMessage;
 import com.pokerogue.helper.global.exception.GlobalCustomException;
 import com.pokerogue.helper.pokemon.data.Pokemon;
-import com.pokerogue.helper.type.collection.TypeMatching;
+import com.pokerogue.helper.type.data.TypeMatching;
 import com.pokerogue.helper.type.data.Type;
 import com.pokerogue.helper.type.repository.TypeMatchingRepository;
 import java.math.BigDecimal;
