+++ backend/pokerogue/src/test/java/com/pokerogue/helper/type/repository/TypeMatchingRepositoryTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.pokerogue.environment.repository.MongoRepositoryTest;
-import com.pokerogue.helper.type.collection.TypeMatching;
+import com.pokerogue.helper.type.data.TypeMatching;
 import com.pokerogue.helper.type.data.Type;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
